+++ src/frontier_util.c
@@ -652,7 +652,7 @@ static const struct WindowTemplate sRankingHallRecordsWindowTemplate =
 };
 
 // Second field - whether the character is female.
-static const u8 sFrontierBrainObjEventGfx[NUM_FRONTIER_FACILITIES][2] =
+static const u16 sFrontierBrainObjEventGfx[NUM_FRONTIER_FACILITIES][2] =
 {
     [FRONTIER_FACILITY_TOWER]   = {OBJ_EVENT_GFX_HOENN_ANABEL,  TRUE},
     [FRONTIER_FACILITY_DOME]    = {OBJ_EVENT_GFX_HOENN_TUCKER,  FALSE},
@@ -2397,6 +2397,7 @@ void CopyFrontierBrainTrainerName(u8 *dst)
 bool8 IsFrontierBrainFemale(void)
 {
     s32 facility = VarGet(VAR_FRONTIER_FACILITY);
+
     return sFrontierBrainObjEventGfx[facility][1];
 }
